Admin, lobby, login, logout implementations
Wow- lots of stuff.  Added all the models to the admin page so you can
create test games really easily.  Changed the __eq__ function names to
compare in the models.py (the __eq__ override was conflicting with
default django functionality).  Added some settings in settings.py.
Minor changes to index.html.  Added a lobby view and template that
renders the currently logged in users games.  Added a login and logout
page for users to login and logout.